Move datafusion-cli to own crate
