+++ Cargo.toml
@@ -18,6 +18,7 @@
 [workspace]
 members = [
     "datafusion",
+    "datafusion-cli",
     "datafusion-examples",
 	"benchmarks",
     "ballista/rust/client",
